requirements.txt
datasets==2.19.1
langchain==0.2.1
langchain_openai==0.1.7
langchain_pinecone==0.1.1
openai==1.30.3
pandas==2.2.2
pinecone==4.0.0
pytest==8.2.1
python-dotenv==1.0.1
ragas==0.1.8
